Modernize gem setup
Modernize the gem structure and setup to match that of fresh gem

    $ bundle gem spectacles

including an updated license, and new changelog and code of conduct.

Also use the conventional Mintest test/ directory and tests named with
the "_test" suffix.
