+++ .gitignore
@@ -1,9 +1,10 @@
-.DS_Store
-*.gem
-.bundle
-.rvmrc
-Gemfile.lock
-pkg/*
-specs/*.db
-coverage
-gemfiles/
+/Gemfile.lock
+/.bundle/
+/.yardoc
+/_yardoc/
+/coverage/
+/doc/
+/pkg/
+/spec/reports/
+/test/*.db
+/tmp/
