+++ CHANGELOG.md
@@ -1,0 +1,53 @@
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+- Add Rails 7.0 support
+- Add support for dumping views from multiple schemas in Postgres
+- Require Ruby 2.7+
+- Require Rails 7.0
+
+## [6.0.0] – 2022-02-21
+
+- Add Rails 6.0 support
+
+## [2.0.0] – 2019-12-17
+
+- Add support for materialized views
+- Drop support for the the mysql gem
+- Require Ruby 2.2+
+
+## [1.2.0] – 2017-07-21
+
+- Add support for skipping views when dumping the schema
+
+## [1.1.0] – 2016-01-15
+
+## [1.0.1] – 2015-09-29
+
+## [1.0.0] – 2015-09-28
+
+## [0.5.3] – 2015-03-16
+
+## [0.5.2] – 2014-12-05
+
+## [0.5.1] – 2014-02-07
+
+## [0.5.0] – 2014-01-03
+
+## [0.4.1] – 2013-11-14
+
+## [0.4.0] – 2013-10-10
+
+## [0.3.1] – 2013-10-04
+
+## [0.3.0] – 2013-09-27
+
+## [0.2.0] – 2013-04-22
+
+## [0.1.0] – 2013-01-14
+
+## [0.0.2] – 2012-02-10
