+++ test/adapters/mysql2_adapter_test.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "test_helper"
 
 describe "Spectacles::SchemaStatements::Mysql2Adapter" do
   config = {
