+++ test/adapters/postgresql_adapter_test.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "test_helper"
 
 describe "Spectacles::SchemaStatements::PostgreSQLAdapter" do
   config = {
