+++ test/spectacles/view_test.rb
@@ -1,4 +1,4 @@
-require "spec_helper"
+require "test_helper"
 
 describe Spectacles::View do
   it "is an abstract class" do
