+++ test/support/minitest/shared_examples.rb
