+++ test/support/view_examples.rb
